requirements.txt
streamlit
pandas
qrcode
pillow
altair
openpyxl